Only prevent updates to the root window
Since we're mainly upset about steam setting the cursor for the root window, I
figured it would be a good idea to only prevent those cursor settings (using
the browser parts of the interface without the cursor changing over buttons is
a little strange).

However, it's doing more work than I thought. It changes the cursor constantly
for whatever window the cursor is currently over (well, the steam window or the
root window, depending on position). In fact, when you're over the steam
window, waving the mouse causes a flurry of cursor changes (and a little cpu
usage bump).

I really wish it would just define the cursor over buttons and undefine the
cursor when it leaves them. I suspect this is to cover up weak coupling between
steam and the windowed elements (browser, flash, etc) but it's still
disappointing.