Use XUndefineCursor when appropriate
Steam still assumes that XC_arrow is the default cursor it should have when
nothing context-specific is being hovered over. Instead, it should inherit the
default cursor by using XUndefineCursor. Since the cursors used by
XDefineCursor seem to get re-used, we can snag the XID of the XC_arrow cursor
and secretly use XUndefineCursor instead when Steam sets the cursor to it. This
preserves both context-sensitive cursors and the default cursor from the root
window.